requirements.txt
mysql-client
Pillow
djangorestframework
djangorestframework-jsonapi
django-filter